Addition of simple Tic Tac Toe
(and yes we call it noughts and crosses in the UK). This bit can't
really be included as part of the progvember idea but it seems
reasonable to include it as some of the ideas are similar and it was the
inspiration. It's probably ever so slightly busted, but oh well.